refactor(page-detection): split provider helper modules
